refactor: 리뷰 수정 버튼 분리
[#480]
